requirements.txt
motor==2.5.1
uvicorn
fastapi
fastapi-simple-cache
httpx
pytest
anyio
pytest-asyncio
pytest-tornasync
pytest-trio
sqlalchemy==2.0.36
mysqlclient==2.2.5
PyMySQL==1.1.1
python-multipart==0.0.17
passlib==1.7.4
PyJWT==2.9.0
yfinance==0.2.48






